% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdf.R
\name{asdf}
\alias{asdf}
\title{Converting arrays to tibbles with or without dimnames}
\usage{
asdf(x, responseName = "n", add_dimnames = FALSE, retval = c("tibble",
  "df"), ...)
}
\arguments{
\item{x}{array}

\item{responseName}{character, name of the variable containing cell values}

\item{add_dimnames}{logical, should dimnames be appended to the result}

\item{retval}{character, class of returned result, \code{tibble} or \code{df} for a data frame}

\item{...}{other arguments passed to \code{\link{provideDimnames}}}
}
\value{
A tibble or data frame, as determined by \code{retval}.
}
\description{
This function returns a "tidy" version of an array: a row per cell with first
variables identifying the dimensions.
}

