% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouter.R
\name{mouter}
\alias{mouter}
\title{A generalization of base::outer()}
\usage{
mouter(..., retval = c("df", "array"), freq_var = "n")
}
\arguments{
\item{...}{vectors or matrices}

\item{retval}{character, whether to return a data frame (\code{"df"}) or
array (\code{"array"})}

\item{freq_var}{character scalar, when returning a data frame this will be
  the name of the variable containing cell values

Given the list of vectors and/or arrays \code{\link{outer}} is used to
combine the elements one-by-one (using \code{\link{Reduce}}).}
}
\value{
Depending on the value of \code{retval}, an array or a data frame with the
outer product of elements supplied through \code{...}.
}
\description{
Given a list of vector or arrays applie \code{\link{outer}} consecutively.
}
\examples{
# As an array
a <- mouter(
  A = c(a=1, b=2) / 3,
  M = matrix(1:4 / 10, 2, 2, dimnames=list(D=letters[5:6], E=letters[7:8])),
  B = c(c=2, d=3) / 5,
  retval="a"
)

str(a)

# Data frame
mouter(
  A = c(a=1, b=2) / 3,
  M = matrix(1:4 / 10, 2, 2, dimnames=list(D=letters[5:6], E=letters[7:8])),
  B = c(c=2, d=3) / 5,
  retval="df"
)
}

